Rename and restructure files
