+++ lib/current_race_screen/new_race_dialog.dart
